+++ CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### Added
+
+- Add endpoint to return crop insurance commodities. [#368](https://github.com/policy-design-lab/pdl-api/issues/368)
+
 ## [1.1.0] - 2026-04-09
 
 ### Added